datasets are now in int form

+++ matrix_dynamic.h
@@ -6,7 +6,7 @@
 #include <assert.h>
 #define FORMAT "%8.3lf"
 #define FORMAT2 "%8d"
-typedef double T;
+typedef int T;
 
 typedef struct {
   int  row_dim, col_dim;
